requirements.txt
attrs==21.2.0
numpy==1.21.2
protobuf==3.19.1
pyzmq==22.3.0
